snowflake-etl/requirements.snowflake.txt
snowflake-snowpark-python
requests